remove lookahead on daily tasks, use correct add model for recurringtasks, deny scheduling recurringtask multiple times, use local date for scheduling

+++ src/TaskPlanner/TaskPlanner.Web/Db/RecurringTaskRepository.cs
@@ -65,7 +65,7 @@ SELECT
 FROM [dbo].[RecurringTask]
 WHERE
     NOT EXISTS (SELECT ScheduledTask.Id FROM ScheduledTask WHERE ScheduledTask.RecurringTaskId = RecurringTask.Id AND ScheduledTask.State = @State)
-", new { State = (int)ScheduledTaskState.Todo });
+", new { State = ScheduledTaskState.Todo });
                 return results.ToList();
             }
         }
